+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.1 [2018-08-01]
 
 - Remove `upgrade` command.
+- Remove `onlyWebpack` option.
 - Remove `esLintOptions, esLintFix, styleLintOptions, styleLintFix, prettierOptions` option.
 - Remove `esLint, styleLint, prettier` option.
 - Remove `eslint,stylelint,prettier` command.